feat: clicking a card adds the id of that pokemon to the state 'selectedPokemonIds', a hashset. if that pokemon id has been selected before, then an alert is shown to indicate that the player has lost the game. the cards still shuffle even if user loses.
a hashset is used to store the state instead of an array because searching a set is done in constant time (highly efficient). the handler for card selection sits in gameboard component and is passed as a prop into card component, as the state for selected pokemon ids sits in gameboard component.