feat: once the player selects a card, that card is snuck into the regenerated set of cards in shuffled order. at the most, there will be 5 out of 10 cards, randomly selected from the list of pokemons already selected, shown to the user in a randomised order. when there are less than 5 cards chosen, those cards are snuck in anyway, again, in shuffled order
use of helper function to shuffle elements and then randomly select elements using that shuffle elements helper. the shuffling of elements implements schwartzian transform. this is sufficient for our purposes as we are unlikely to have more than hundreds of pokemon cards selected by the user, so this is more readable than fischer yates shuffle