feat: ensure responsiveness on smaller screens. refactor: use css modules.
on narrow screens, everything is resized. at a certain breakpoint (super narrow mobile screens), we use two columns of 5 rows to better use the limited page space. even narrower portrait will remove the name of the pokemons entirely. css modules improves maintainability by preventing name clashing when scaling application, and also improves performance (better than using css-in-js alternative, or css libraries). fixes #11. note: there is some 'jumping' in the positioning of the header during loading states. i've hacked together a bunch of media queries to fix this as i don't know how to ensure the size of the main section, or the loadingScreen component outer element with css. i don't want to be generating placeholder Card components unnecessarily as this is not performant, but the <main> section essentially needs to be stretched out