test: add tests for components to check rendering and expected behaviour. docs: update README. chore: move CSS modules into folders with components.
having tests and CSS modules in the same folder as the components increases ease of maintainability. this ticket fixes #15, although the subtasks will be completed separately: #25 will test the hook, which has some issues that need to be resolved. #26 will implement e2e testing. #10 is tested here, along with ensuring the helper function to randomly select Pokemon IDs do not return any ids that are duplicates (that would be bad for gameplay). README was updated to increase size of the logo, update the screenshot to have hover effect apparent, and added more explanation around using Vitest and RTL. Removed acknowledgement for an old loading image (Pokeball) as we are now just using the Masterball that Muhammad Jazman created.